2011-01-13  Basile Starynkevitch  <basile@starynkevitch.net>
	* melt-runtime.c (meltgc_add_ourt_cstr_len): Better signed int
	compare to avoid warning.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/melt-branch@168746 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/melt-runtime.c
@@ -3194,7 +3194,7 @@ meltgc_add_out_cstr_len (melt_ptr_t outbuf_p, const char *str, int slen)
     return;
   if (slen<0) 
     slen = strlen(str);
-  if (slen<sizeof(tinybuf)-1) {
+  if (slen<(int) sizeof(tinybuf)-1) {
     memset (tinybuf, 0, sizeof(tinybuf));
     memcpy (tinybuf, str, slen);
     dupstr = tinybuf;
